requirements.txt
Flask>=2.3.0
requests>=2.28.0
dnspython>=2.3.0
python-whois>=0.8.0
urllib3>=1.26.0
concurrent-futures>=3.1.1
argparse>=1.4.0
Jinja2>=3.1.0
Werkzeug>=2.3.0
MarkupSafe>=2.1.0
